feat: Allow libeuicc to be installed as a dynamic library
...based on option LPAC_DYNAMIC_LIBEUICC, defaults to OFF.

This makes it possible to use libeuicc on Linux without submoduling.

Note that this would require any future breaking change to increment the
MAJOR version of the project in the main CMakeLists.txt. This version
number is now a public interface and must be maintained.

As part of this commit, I have modified the version code in CMakeLists
to match the current release version.