driver: add uqmi backend
This commit adds a backend to interface with an eUICC connected to a
compatible modem using the uqmi application.

This allows OpenWrt to manage an eUICC including the download and
installation of SIM profiles.

This was previously not possible with most modems, as the APDU
operations were aborted due to exceeding the timeout imposed
on the AT interface by the modem firmware.

Tested-on: Quectel EC25 / Quectel EP06 / Quectel RG520N

Signed-off-by: David Bauer <david.bauer@uniberg.com>