driver: apdu: mbim: sleep while waiting for sim
When waiting for the sim to become available after potentially
switching, we don't actually do much waiting. We're basically only
delayed by the speed of the communication here. If the switch doesn't
happen quickly enough, this may result in us deciding a sim didn't
become available, but subsequent calls will produce usable results.

So introduce a small 50 ms wait time, so we limit the total waiting
time to about 1 second. This I think should be enough for the sim to
actually become available.

resolve #241

Signed-off-by: Frans Klaver <frans.klaver@vislink.com>