refactor(lpac): use __attribute__((cleanup)) to do SBRM (#270)
* refactor(driver): make json_print() reusable

So we can get rid of redefinition issue of json_print() and use helper.h
in all driver code.

Signed-off-by: Coelacanthus <uwu@coelacanthus.name>

* refactor(driver): use __attribute__((cleanup)) to do SBRM

Signed-off-by: Coelacanthus <uwu@coelacanthus.name>

* refactor: split helpers of driver into a new object library

* To use in lpac frontend code.
* To add lpac path prefix to avoid common header name conflict.

Signed-off-by: Coelacanthus <uwu@coelacanthus.name>

* chore(gitignore): add clangd per-project index directory

Signed-off-by: Coelacanthus <uwu@coelacanthus.name>

* refactor(src): use __attribute__((cleanup)) to do SBRM

Signed-off-by: Coelacanthus <uwu@coelacanthus.name>

* fix(applet/notification/replay): tidy #include usage

Signed-off-by: Coelacanthus <uwu@coelacanthus.name>

* fix(applet/profile/download): memory leak of error_detail

In current code, we always do strdup() on error_detail to avoid invoking
free() on static string. It's safe but will cause memory leak. Let's do
strdup() when assign static string to error_detail to avoid memory leak
and still keep safe.

Signed-off-by: Coelacanthus <uwu@coelacanthus.name>

* fix(mingw): get rid of MinGW's stupid interface macro definition

In [1], they defined interface to struct although the code is not
Object-C or Object-C++. It's stupid and violate C/C++ standard: all
identifiers used by compiler or standard library should either be
a keyword, or be start with an underline plus a uppercase latter or two
underline to avoid conflict with user defined identifiers.

[1]: https://github.com/mingw-w64/mingw-w64/commit/a51690d44ec4f03dcf5bfb1f3573eb0d720b3c12

Signed-off-by: Coelacanthus <uwu@coelacanthus.name>

* refactor(lpac): remove unnecessary goto logic

Signed-off-by: Coelacanthus <uwu@coelacanthus.name>

---------

Signed-off-by: Coelacanthus <uwu@coelacanthus.name>